Refactored code and added unit tests
Split up code to allow unit testing
Added folder unit_tests and a makefile
Refactored the code and added Makefile and unittests, however the makefile is not working.
Removed generated object files
Moved the extern definition from the main file & fixed some warnings

The extern was correctly in the header file, but the definition of it
was provided in the _main file rather than the actual definition file.
This meant that the unit test linked not to the _main.c file would not
have a definition of the extern.